Module Rfsm.Compiler

The command-line compiler

module type T = sig ... end

Output signature of the functor Compiler.Make

module type PARSER = sig ... end

Signature for the Parser input to the functor Compiler.Make

module type LEXER = sig ... end

Signature for the Lexer input to the functor Compiler.Make

module Make (L : Host.T) (Lexer : LEXER) (Parser : PARSER with type token = Lexer.token and type program = L.Syntax.program and type fragment_obj = L.Syntax.fragment_obj) : T

Functor building the compiler structure given a language definition, a lexer and a parser implementation