Module Systemc.G

module Static : Guest.STATIC

module Syntax : Guest.SYNTAX

type value

val pp_typed_symbol : Stdlib.Format.formatter -> (Ident.t * Syntax.type_expr) -> unit

pp_typed_symbol fmt (id,ty) should print, on formatter fmt, the SystemC declaration of symbol id with type ty.

val pp_type_expr : Stdlib.Format.formatter -> Syntax.type_expr -> unit

pp_type_expr fmt te should print, on formatter fmt, the SystemC type corresponding to type expression te.

val pp_type_decl : Stdlib.Format.formatter -> Syntax.type_decl -> unit

val pp_type_impl : Stdlib.Format.formatter -> Syntax.type_decl -> unit

pp_type_decl fmt td (resp. pp_type_impl fmt td) should print, on formatter fmt, the SystemC declaration and implementation (in the .h and .cpp file resp.) corresponding to type declaration td.

val pp_expr : Stdlib.Format.formatter -> Syntax.expr -> unit

pp_expr fmt e should print, on formatter fmt, the SystemC expression corresponding to expression e.

val pp_typ : Stdlib.Format.formatter -> Syntax.Types.typ -> unit

pp_typ fmt ty should print, on formatter fmt, the SystemC type corresponding to type ty.

val pp_lval : Stdlib.Format.formatter -> Syntax.lval -> unit

pp_lval fmt l should print, on formatter fmt, the SystemC representation of l-value l.

val pp_value : Stdlib.Format.formatter -> value -> unit

pp_value fmt v should print, on formatter fmt, the SystemC value corresponding to value v.