Module Rfsm.Depg

Dependency graphs

module G : sig ... end

Graphs for which vertices are simple names

module M : sig ... end

Graphs for which vertices are simple names

module TS : sig ... end

For mapping names to vertices and to full descriptors

The signature for graph nodes

module type NODE = sig ... end

The signature of the module performing dependency sorting of a graph of nodes

module type T = sig ... end

The functor building such a module

module Make (N : NODE) : T with type node = N.t and type context = N.context