Module type Depg.NODE

type t

type context

val name_of : t -> Ident.t

name_of n should return a unique name for node n

val depends_on : context -> t -> t -> bool

name_of n should return a unique name for node n

depends_on c n n' returns true if node n depends on node n' in context c, false otherwise.