Module Rfsm.Evseq

Event sequences

An event sequence is a sequence of dated event sets. Ex: [{H,x<-1}@t=10; {H}@t=20, ...]. Event sequences are to describe both stimuli and responses.

module type EVSEQ = sig ... end

module Make (ES : Evset.T) : EVSEQ with module Evset = ES and type value = ES.Event.Value.t