Module T.Ctask

module Syntax = Syntax

val pp_typed_symbol : Stdlib.Format.formatter -> (Ident.t * Syntax.type_expr) -> unit

pp_typed_symbol fmt (id,ty) should print, on formatter fmt, the C declaration of symbol id with type ty. For example, for an int named "x", this should be "int x", and for an array of 8 ints named "a", this should be "int a[8]".

val pp_type_expr : Stdlib.Format.formatter -> Syntax.type_expr -> unit

pp_type_expr fmt te should print, on formatter fmt, the C type corresponding to type expression te.

val pp_type_decl : Stdlib.Format.formatter -> Syntax.type_decl -> unit

pp_type_decl fmt td should print, on formatter fmt, the C statement corresponding to type declaration td.

val pp_expr : Stdlib.Format.formatter -> Syntax.expr -> unit

pp_expr fmt e should print, on formatter fmt, the C expression corresponding to expression e.