Module Guest.Types

type typ

Guest-level types

Constructors

val no_type : typ

Special value denoting an undefined type

val mk_type_fun : typ list -> typ -> typ

mk_type_fun [t1;...;tn] t should return the type of a function taking n arguments with type t1, ..., tn and returning a result with type t.

Inspectors

val is_event_type : typ -> bool

val is_bool_type : typ -> bool

is_event_type t (resp. is_bool_type t) should return true iff type t is the event (resp. boolean) type. Returns false if the guest language has no such type.

Printing

val pp_typ : ?abbrev:bool -> Stdlib.Format.formatter -> typ -> unit

pp_type abbrev fmt t should print a readable representation of type t on formatter fmt. The optional abbrev argument can be used to require an abbreviated form (for example, for a record, to print only the name of the record, without the description of the fields.