Module Rfsm__Dynamic

Dynamic models (used by the simulator)

FSM

type fsm = {

f_static : Rfsm.Fsm.inst;
Static representation

f_vars : (string * (Rfsm.Types.typ * Rfsm.Expr.value)) list;
name, (type, value)

f_state : string;
Current state

f_has_reacted : bool;
true when implied in the last reaction

}

Local evaluation environment

type lenv = (string * Rfsm.Expr.value) list

Global evaluation environment

type genv = {

fe_inputs : (string * (Rfsm.Types.typ * Rfsm.Expr.value)) list;
Global inputs

fe_csts : (string * (Rfsm.Types.typ * Rfsm.Expr.value)) list;
Global constants

fe_fns : (string * (Rfsm.Types.typ * Rfsm.Expr.value)) list;
Global functions

fe_vars : (string * (Rfsm.Types.typ * Rfsm.Expr.value)) list;
Shared variables

fe_evs : (string * (Rfsm.Types.typ * Rfsm.Expr.value)) list;
Shared events

}

Builders

val make_fsm : Rfsm.Fsm.inst -> fsm

create sf creates a dynamic FSM instance from a static description sf

type event = loc * Rfsm.Expr.value

Event location, new value

and loc =

| LVar of Rfsm.Ident.t
Scalar

| LArrInd of Rfsm.Ident.t * int
1D array location

| LRField of Rfsm.Ident.t * string
Record field

Exceptions

exception IllegalTrans of fsm * string

exception IllegalAction of fsm * Rfsm.Action.t

exception Undeterminate of fsm * string * Rfsm.Types.date

exception NonDetTrans of fsm * Rfsm.Fsm.transition list * Rfsm.Types.date

exception NonAtomicIoWrite of fsm * Rfsm.Action.t

Accessors

val fireable : fsm -> lenv -> Rfsm.Fsm.Repr.transition -> bool

fireable f env t returns true iff transition t in FSM f is fireable, given ocal env env.

val check_cond : fsm -> lenv -> Rfsm.Condition.t -> bool

check_cond f env c evaluatues condition c in FSM f, in the context of local env env.

val is_event_set : lenv -> Rfsm.Condition.event -> bool

is_event_set env ev indicates whether event ev is set in local environment env

Simulation

val init : genv -> fsm -> fsm * event list

init env f performs the initial transition of FSM f, in global environment env, Returns an updated fsm and list of events consisting of

updates to global outputs or shared objects

updates to local variables (including state move)

val react : Rfsm.Types.date -> genv -> fsm -> fsm * event list

react t env f compute the reaction, at time t of FSM f in global environment env. The global environment contains the values of global inputs, shared objects and global functions/constants. As for init, returns an updated fsm and list of events.