Module Rfsm__Eval

Evaluating and manipulating expressions

exception Unknown_id of string

exception Illegal_expr of Rfsm.Expr.t

exception Illegal_application of Rfsm.Expr.t

exception Illegal_array_access of Rfsm.Expr.t

exception Illegal_bit_range_access of Rfsm.Expr.t

exception Invalid_array_access of string * int

exception Illegal_record_access of Rfsm.Expr.t

exception Non_static_expr of Rfsm.Expr.t * Rfsm.Expr.t

type env = (string * Rfsm.Expr.value) list

val subst : (string * Rfsm.Expr.value) list -> Rfsm.Expr.t -> Rfsm.Expr.t

subst senv e substitutes each occurence of variable v listed in senv by its value in e

val eval : (string * Rfsm.Expr.value) list -> Rfsm.Expr.t -> Rfsm.Expr.value

eval env e evaluates expression e in the context of environment env