Module Rfsm__Typing

Typing

type tenv = {

te_vars : (string * Rfsm.Types.typ) list;

te_ctors : (string * Rfsm.Types.typ) list;

te_rfields : (string * Rfsm.Types.typ) list;

te_defns : (string * Rfsm.Types.typ) list;

te_prims : (string * Rfsm.Types.typ_scheme) list;

}

val builtin_tenv : tenv

Exceptions

exception Undef_symbol of string * string * string

where, what, name

exception Type_error of Rfsm.Expr.t * Rfsm.Types.typ * Rfsm.Types.typ

where, what, name

exception Typing_error of string * string * Rfsm.Types.typ * Rfsm.Types.typ

what, where, type, type

exception Illegal_cast of Rfsm.Expr.t

what, where, type, type

exception Unbound_type_ctor of string

exception Invalid_record_access of Rfsm.Expr.t

Typing

val type_of_type_expr : tenv -> Rfsm.Type_expr.t -> Rfsm.Types.typ

type_expression env te returns the type represented by type expression te in environment env

val type_expression : tenv -> Rfsm.Expr.t -> Rfsm.Types.typ

type_expression env e returns the type of expression e in environment env, performing all required type checks.

val type_check_stim : tenv -> string -> Rfsm.Types.typ -> Rfsm.Global.stim_desc -> unit

type_stimuli env id ty sd type checks an input stimuli description sd against type ty

val type_fsm_model : tenv -> Rfsm.Fsm.model -> unit

type_fsm_model env m type checks an FSM model in environment env

val type_fsm_inst : tenv -> Rfsm.Fsm.inst -> unit

type_fsm_inst env f type checks an FSM instance f in environment env.

Printers

val dump_tenv : Stdlib.out_channel -> tenv -> unit